+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/SPORatifySpec.hs
@@ -20,11 +20,14 @@ import Cardano.Ledger.Conway.Governance (
   RatifyEnv (..),
   RatifyState,
   Vote (..),
+  VotingStakePoolDistr,
   ensProtVerL,
   gasAction,
   gasActionL,
+  mkVotingStakePoolDistr,
   rsEnactStateL,
   votingStakePoolThreshold,
+  vspdTotalVotingStakeL,
  )
 import Cardano.Ledger.Conway.Rules (
   spoAccepted,
@@ -66,7 +69,7 @@ acceptedRatioProp = do
               actual =
                 spoAcceptedRatio @era
                   re
-                    { reStakePoolDistr = distr
+                    { reVotingStakePoolDistr = distr
                     , reAccounts = accountsFromDelegatees delegatees
                     , reStakePools = stakePools
                     }
@@ -99,7 +102,7 @@ noStakeProp =
   prop @((RatifyEnv era, RatifyState era, GovActionState era) -> IO ())
     "If there is no stake, accept iff threshold is zero"
     ( \(re, rs, gas) ->
-        let re' = re {reStakePoolDistr = PoolDistr Map.empty (knownNonZeroCoin @100)}
+        let re' = re {reVotingStakePoolDistr = mkVotingStakePoolDistr mempty (knownNonZeroCoin @100)}
          in spoAccepted @era re' rs gas
               `shouldBe` (votingStakePoolThreshold @era rs (gasAction gas) == SJust minBound)
     )
@@ -116,7 +119,7 @@ allAbstainProp =
         spoAcceptedRatio
           @era
           re
-            { reStakePoolDistr = distr
+            { reVotingStakePoolDistr = distr
             , reAccounts = accountsFromDelegatees delegatees
             , reStakePools = stakePools
             }
@@ -135,7 +138,7 @@ noVotesProp =
       $ \TestData {..} ->
         spoAcceptedRatio
           @era
-          re {reStakePoolDistr = distr}
+          re {reVotingStakePoolDistr = distr}
           gas {gasStakePoolVotes = votes}
           (rs ^. rsEnactStateL . ensProtVerL)
           `shouldBe` 0
@@ -153,7 +156,7 @@ allYesProp =
             let acceptedRatio =
                   spoAcceptedRatio
                     @era
-                    re {reStakePoolDistr = distr}
+                    re {reVotingStakePoolDistr = distr}
                     gas {gasStakePoolVotes = votes}
                     (rs ^. rsEnactStateL . ensProtVerL)
              in acceptedRatio `shouldBe` 1
@@ -170,13 +173,13 @@ noConfidenceProp =
       $ \TestData {..} ->
         spoAcceptedRatio
           @era
-          re {reStakePoolDistr = distr}
+          re {reVotingStakePoolDistr = distr}
           gas {gasStakePoolVotes = votes}
           (rs ^. rsEnactStateL . ensProtVerL)
           `shouldBe` 0
 
 data TestData era = TestData
-  { distr :: PoolDistr
+  { distr :: VotingStakePoolDistr
   , votes :: Map (KeyHash StakePool) Vote
   , totalStake :: NonZero Coin
   , stakeYes :: Coin
@@ -212,19 +215,17 @@ genTestData Ratios {yes, no, abstain, alwaysAbstain, noConfidence} = do
     (poolsYes, poolsNo, poolsAbstain, poolsAlwaysAbstain, poolsNoConfidence, rest) =
       splitByPct yes no abstain alwaysAbstain noConfidence pools
   distr <- do
-    vrf <- arbitrary
-    bls <- arbitrary
     let
-      indivStake = IndividualPoolStake (1 % unCoin (unNonZero totalStake)) (CompactCoin 1) vrf bls
+      stakeAmount = CompactCoin 1
       distr =
         unionAllFromLists
-          [ (poolsYes, indivStake)
-          , (poolsNo, indivStake)
-          , (poolsAbstain, indivStake)
-          , (poolsAlwaysAbstain, indivStake)
-          , (poolsNoConfidence, indivStake)
+          [ (poolsYes, stakeAmount)
+          , (poolsNo, stakeAmount)
+          , (poolsAbstain, stakeAmount)
+          , (poolsAlwaysAbstain, stakeAmount)
+          , (poolsNoConfidence, stakeAmount)
           ]
-    pure $ PoolDistr distr totalStake
+    pure $ mkVotingStakePoolDistr distr totalStake
 
   poolStateAA <- genPoolState poolsAlwaysAbstain
   poolStateNC <- genPoolState poolsNoConfidence
@@ -237,7 +238,7 @@ genTestData Ratios {yes, no, abstain, alwaysAbstain, noConfidence} = do
     TestData
       { distr
       , votes
-      , totalStake = pdTotalActiveStake distr
+      , totalStake = distr ^. vspdTotalVotingStakeL
       , stakeYes = Coin . fromIntegral $ length poolsYes
       , stakeNo = Coin . fromIntegral $ length poolsNo
       , stakeAbstain = Coin . fromIntegral $ length poolsAbstain